Split a bit more cli/tests/e2e
test plan:
cd cli
pipenv run pytest
