+++ cli/tests/e2e-pro/test_config_resolver.py
