+++ cli/tests/e2e-pysemgrep/test_utility_commands.py
@@ -4,15 +4,7 @@ import subprocess
 import pytest
 from tests.semgrep_runner import SEMGREP_BASE_SCAN_COMMAND
 
-
-@pytest.mark.kinda_slow
-def test_version():
-    result = subprocess.check_output(
-        SEMGREP_BASE_SCAN_COMMAND + ["--version", "--disable-version-check"],
-        encoding="utf-8",
-    )
-
-    assert re.match(r"\d+\.\d+\.\d+", result)
+# used to be in e2e/test_utility_commands.py
 
 
 @pytest.mark.kinda_slow
