+++ cli/tests/e2e/snapshots/test_cli_test_fixtest/test_fixtest_test2_no_json/output.txt
