+++ cli/tests/e2e/snapshots/test_cli_test_fixtest/test_fixtest_test3_no_json/output.txt
