+++ cli/tests/e2e/snapshots/test_cli_test_fixtest/test_fixtest_test4_no_json/results.txt
