+++ cli/tests/e2e/snapshots/test_extract/test_extract/results.json
