+++ cli/tests/e2e/snapshots/test_extract/test_extract_exclude/results.json
