+++ cli/tests/e2e/snapshots/test_output_sarif/test_sarif_output_include_nosemgrep/results.sarif
