+++ cli/tests/e2e/snapshots/test_output_sarif/test_sarif_output_with_autofix/results.sarif
