+++ cli/tests/e2e/snapshots/test_sca/test_sca_lockfile_only_output/results.txt
