+++ cli/tests/e2e/snapshots/test_sca/test_sca_output/results.txt
