+++ cli/tests/e2e/snapshots/test_secret/test_cli_test_secret_rule/results.txt
