+++ cli/tests/e2e/test_cli_test_fixtest.py
